feat: MVP code for bonking given process
Creates a loop that will periodically check the CPU usage of a
process. If the process uses more than 90% CPU for five minutes we send
a sigkill signal to force quit the process.

The code will also find the process pid first and store it so it does
not need to look it up everytime it needs to check its CPU usage.

TODO: Build application with GitHub actions, add brew setup, etc.